Microservices: add services loans, cards, accounts

+++ Microservices (H2, JPA, WEB, Lombok, Validation, Devtools) - Building Microservice/accounts/src/main/java/com/didan/microservices/accounts/entity/Accounts.java
@@ -16,7 +16,7 @@ import lombok.ToString;
 public class Accounts extends BasicClass{
 
 	@Column(name = "customer_id")
-	private int customerId;
+	private Long customerId;
 	
 	@Id
 	@Column(name = "account_number")
